perf: add confirmation alert when user delete note

+++ src/app/modules/note/pages/note-overview-page/note-overview-page.component.ts
@@ -37,6 +37,10 @@ export class NoteOverviewPageComponent implements OnInit {
   }
 
   onNoteDelete(note: Note): void {
-    this.noteStoreFacade.deleteNote(note);
+    const confirmed = confirm('Are you sure you want to delete this note?');
+
+    if (confirmed) {
+      this.noteStoreFacade.deleteNote(note);
+    }
   }
 }
